api\server: Factored out UnixHttp on Windows, supported on Linux only.
